Share mobile WarlockApp; rework settings nav; chromeless game screen
Extract a single WarlockApp into compose/mobileMain used by both Android
and iOS (deleting the duplicated androidApp WarlockApp and the inline
IosWarlockApp). The dashboard keeps its top app bar and drawer, but the
drawer now points to a single Settings entry instead of one item per
section.

Add a top-level SettingsScreen that lists the setting categories and
navigates into each via the existing SettingsContent.

The connected game screen is now chromeless (no top bar); a settings
button and a window-list toggle live side-by-side under the entry bar in
the game bottom bar. Window-list visibility is local GameView state,
replacing the previously hardcoded sideBarVisible = false.
