+++ compose/src/iosMain/kotlin/warlockfe/warlock3/compose/MainViewController.kt
@@ -1,47 +1,16 @@
 package warlockfe.warlock3.compose
 
-import androidx.compose.foundation.isSystemInDarkTheme
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.DrawerValue
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.ModalDrawerSheet
-import androidx.compose.material3.ModalNavigationDrawer
-import androidx.compose.material3.NavigationDrawerItem
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.rememberDrawerState
 import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.ComposeUIViewController
 import co.touchlab.kermit.Logger
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
-import kotlinx.coroutines.launch
-import org.jetbrains.compose.resources.painterResource
-import warlockfe.warlock3.compose.components.ScrollableColumn
 import warlockfe.warlock3.compose.generated.resources.Res
-import warlockfe.warlock3.compose.generated.resources.menu
-import warlockfe.warlock3.compose.model.GameState
-import warlockfe.warlock3.compose.ui.settings.SettingsContent
-import warlockfe.warlock3.compose.ui.settings.SettingsPage
-import warlockfe.warlock3.compose.ui.theme.AppTheme
 import warlockfe.warlock3.compose.util.LocalSkin
 import warlockfe.warlock3.compose.util.SkinLoader
-import warlockfe.warlock3.core.client.GameCharacter
-import warlockfe.warlock3.core.prefs.ThemeSetting
-import warlockfe.warlock3.core.sge.SgeSettings
 
 @Suppress("ktlint:standard:function-naming")
 fun MainViewController() =
@@ -69,116 +38,9 @@ fun MainViewController() =
         CompositionLocalProvider(
             LocalSkin provides skin,
         ) {
-            IosWarlockApp(
+            WarlockApp(
                 appContainer = appContainer,
                 sgeSettings = sgeSettings,
             )
         }
     }
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Suppress("FunctionName")
-@androidx.compose.runtime.Composable
-private fun IosWarlockApp(
-    appContainer: AppContainer,
-    sgeSettings: SgeSettings,
-) {
-    val themeSetting by appContainer.clientSettings.observeTheme().collectAsState(ThemeSetting.AUTO)
-    val darkMode =
-        when (themeSetting) {
-            ThemeSetting.AUTO -> isSystemInDarkTheme()
-            ThemeSetting.LIGHT -> false
-            ThemeSetting.DARK -> true
-        }
-    LaunchedEffect(darkMode) { appContainer.darkMode.value = darkMode }
-    AppTheme(useDarkTheme = darkMode) {
-        val gameState = remember { GameState() }
-        var settingsPage by remember { mutableStateOf<SettingsPage?>(null) }
-        var currentCharacter: GameCharacter? by remember { mutableStateOf(null) }
-
-        val drawerState = rememberDrawerState(DrawerValue.Closed)
-        val scope = rememberCoroutineScope()
-
-        ModalNavigationDrawer(
-            drawerState = drawerState,
-            drawerContent = {
-                ModalDrawerSheet {
-                    ScrollableColumn {
-                        NavigationDrawerItem(
-                            modifier = Modifier.padding(horizontal = 12.dp),
-                            label = { Text("Main") },
-                            onClick = {
-                                scope.launch { drawerState.close() }
-                                settingsPage = null
-                            },
-                            selected = settingsPage == null,
-                        )
-                        SettingsPage.entries.forEach { page ->
-                            NavigationDrawerItem(
-                                modifier = Modifier.padding(horizontal = 12.dp),
-                                label = { Text(page.title) },
-                                onClick = {
-                                    scope.launch { drawerState.close() }
-                                    settingsPage = page
-                                },
-                                selected = settingsPage == page,
-                            )
-                        }
-                    }
-                }
-            },
-        ) {
-            Scaffold(
-                topBar = {
-                    TopAppBar(
-                        title = { Text(text = "Warlock", maxLines = 1) },
-                        navigationIcon = {
-                            IconButton(
-                                onClick = {
-                                    scope.launch { drawerState.open() }
-                                },
-                            ) {
-                                Icon(
-                                    painter = painterResource(Res.drawable.menu),
-                                    contentDescription = null,
-                                )
-                            }
-                        },
-                    )
-                },
-            ) { padding ->
-                Box(Modifier.padding(padding)) {
-                    if (settingsPage == null) {
-                        MainScreen(
-                            sgeViewModelFactory = appContainer.sgeViewModelFactory,
-                            dashboardViewModelFactory = appContainer.dashboardViewModelFactory,
-                            gameState = gameState,
-                            updateCurrentCharacter = { currentCharacter = it },
-                            sgeSettings = sgeSettings,
-                            sideBarVisible = false,
-                        )
-                    } else {
-                        Box(Modifier.padding(16.dp)) {
-                            SettingsContent(
-                                page = settingsPage!!,
-                                currentCharacter = currentCharacter,
-                                variableRepository = appContainer.variableRepository,
-                                macroRepository = appContainer.macroRepository,
-                                presetRepository = appContainer.presetRepository,
-                                highlightRepository = appContainer.highlightRepository,
-                                nameRepository = appContainer.nameRepository,
-                                characterSettingsRepository = appContainer.characterSettingsRepository,
-                                aliasRepository = appContainer.aliasRepository,
-                                scriptDirRepository = appContainer.scriptDirRepository,
-                                alterationRepository = appContainer.alterationRepository,
-                                characterRepository = appContainer.characterRepository,
-                                clientSettingRepository = appContainer.clientSettings,
-                                wraythImporter = appContainer.wraythImporter,
-                            )
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
